+++ smusic/src/main/java/com/bzh/smusic/lib/IMusicListener.kt
@@ -1,5 +1,5 @@
 package com.bzh.smusic.lib
 
 interface IMusicListener {
-    fun onTime(totalTime: Long, currentTime: Long)
+    fun onTime(totalTime: Int, currentTime: Int)
 }
